Fix bug in options choosing code
The `available_choices` function selects the possible responses that
the user is allowed to choose.  It should provide three options:

1. The first unseen option from "Option 1 list"
2. The first unseen option from "Option 2 list"
3. The first unseen option from "Option 3 list"

If the there are no options left in the given list it should fall
back to offering options from the other lists.

The implementation of this logic was broken.  The code

    shift_option(opts, [0, 2, 1])

was intended to select from list 1, falling back to list 3 and then
as a last resort select from list 1.  The logic for iterating over
`[0, 2, 1]` was wrong: instead of using the values in the list we
would use the indices.  In effect we would always count `0`, `1`,
`2`, essentially ignoring the order requested.

This commit fixes that issue.  It shows the value of having tests,
of which this project still has none :(.